+++ src/ethereum/info/zksyncInfo.ts
@@ -97,9 +97,9 @@ const networkInfo: EthereumNetworkInfo = {
     {
       type: 'evmscan',
       servers: [
+        'https://api.etherscan.io',
         'https://block-explorer-api.mainnet.zksync.io',
-        'https://api-era.zksync.network',
-        'https://zksync.blockscout.com/api'
+        'https://api-era.zksync.network'
       ]
     }
   ],